Fix wrong number of args for IsSlice
The following error message is received when the hugo site is run with
this module

<reflect>: wrong number of args for IsSlice: want 1 got 0

In the twitter.html partial, there is an if statement that checks if
.Params.authors is a slice. If it is, the code attempts to retrie the Twitter
handle of the first author using index . 0. However, the if statement is not
properly closed with an end statement. As a result, the subsequent with
statement is being executed regardless of whether .Params.authors is a slice or
not, and the reflect.IsSlice function is being called without any arguments.

To fix this error, added an end statement after the last end statement in the
twitter.html partial. This properly closes the if statement and ensure that the
subsequent code is only executed when .Params.authors is actually a slice: